pyproject.toml
[project]
name = "legalease"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "annotated-types==0.7.0",
    "anthropic==0.86.0",
    "anyio==4.11.0",
    "asttokens==3.0.1",
    "attrs==26.1.0",
    "bcrypt==4.2.0",
    "blinker==1.9.0",
    "certifi==2026.2.25",
    "cffi==2.0.0",
    "charset-normalizer==3.4.6",
    "click==8.3.1",
    "colorama==0.4.6",
    "comm==0.2.3",
    "contourpy==1.3.3",
    "cryptography==46.0.6",
    "cycler==0.12.1",
    "debugpy==1.8.19",
    "decorator==5.2.1",
    "distro==1.9.0",
    "dnspython==2.8.0",
    "docstring-parser==0.17.0",
    "dotenv>=0.9.9",
    "executing==2.2.1",
    "flask==3.1.3",
    "flask-cors==6.0.1",
    "flask-login==0.6.3",
    "flask-mail==0.10.0",
    "fonttools==4.60.1",
    "google-ai-generativelanguage==0.6.15",
    "google-api-core==2.25.2",
    "google-api-python-client==2.193.0",
    "google-auth==2.49.1",
    "google-auth-httplib2==0.3.0",
    "google-genai==1.68.0",
    "google-generativeai>=0.8.6",
    "googleapis-common-protos==1.73.1",
    "greenlet==3.2.4",
    "grpcio==1.78.0",
    "grpcio-status==1.71.2",
    "gunicorn==23.0.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httplib2==0.31.2",
    "httpx==0.28.1",
    "idna==3.11",
    "iniconfig==2.3.0",
    "ipykernel==7.1.0",
    "ipython==9.9.0",
    "ipython-pygments-lexers==1.1.1",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "joblib==1.5.2",
    "jupyter-client==8.8.0",
    "jupyter-core==5.9.1",
    "kafka==1.3.5",
    "kafka-python==2.2.15",
    "kagglehub==0.3.13",
    "kiwisolver==1.4.9",
    "lxml==6.0.2",
    "markupsafe==3.0.3",
    "matplotlib==3.10.7",
    "matplotlib-inline==0.2.1",
    "nest-asyncio==1.6.0",
    "nltk==3.9.2",
    "numpy==2.2.6",
    "obspy==1.4.2",
    "openai==2.7.2",
    "opencv-python==4.12.0.88",
    "outcome==1.3.0.post0",
    "packaging==26.0",
    "pandas==2.3.3",
    "parso==0.8.5",
    "pdfminer-six==20251230",
    "pdfplumber==0.11.9",
    "pexpect==4.9.0",
    "pillow==12.1.1",
    "platformdirs==4.5.1",
    "playwright==1.58.0",
    "pluggy==1.6.0",
    "prompt-toolkit==3.0.52",
    "proto-plus==1.27.2",
    "protobuf==5.29.6",
    "psutil==7.2.1",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.3",
    "pyasn1==0.6.3",
    "pyasn1-modules==0.4.2",
    "pycparser==3.0",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pyee==13.0.1",
    "pygments==2.19.2",
    "pymongo==4.16.0",
    "pyparsing==3.3.2",
    "pypdfium2==5.6.0",
    "pysocks==1.7.1",
    "pytesseract==0.3.13",
    "pytest==9.0.2",
    "python-dateutil==2.9.0.post0",
    "python-dotenv==1.2.2",
    "pytz==2025.2",
    "pyyaml==6.0.3",
    "pyzmq==27.1.0",
    "regex==2026.1.15",
    "requests==2.33.0",
    "scikit-learn==1.7.2",
    "scipy==1.16.2",
    "selenium==4.41.0",
    "setuptools==80.9.0",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sortedcontainers==2.4.0",
    "sqlalchemy==1.4.54",
    "stack-data==0.6.3",
    "tenacity==9.1.4",
    "threadpoolctl==3.6.0",
    "tornado==6.5.4",
    "tqdm==4.67.3",
    "traitlets==5.14.3",
    "trio==0.33.0",
    "trio-websocket==0.12.2",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "tzdata==2025.2",
    "uritemplate==4.2.0",
    "urllib3==2.6.3",
    "wcwidth==0.2.14",
    "webdriver-manager==4.0.2",
    "websocket-client==1.9.0",
    "websockets==16.0",
    "werkzeug==3.1.7",
    "wsproto==1.3.2",
]
